Add logic and tests for MAX_BATCH_SIZE.
This commit adds the logic to only allow a certain
number of metrics when batching metrics in the
metric_publisher_function, along with corresponding unit
tests. It also adds in helper functions to account for
latencies in APIs used throughout the app. This is so that
the integration test can be run without race conditions.
make passed.